requirements.txt
flask
google.generativeai
markdown
langchain
langchain_core
langchain_groq
langchain_community
transformers
pytesseract
textwrap